Hiding Player Model
When you look around, gun is all looking around perfectly but if you look down, the hand of the player model is facing up right there and that's not great.
And of course, when you run along, you can see inside  of  player body and you can see arms and legs moving around.
What I did is do thaw most games do which is when you are in the first person view. So basically when you're the player in control we're going to get rid of the "player model".