Displaying the Correct Weapons
In this commit, I setup the correct animation for putting weapons in player hands and displaying the correct weapons.
If you switch the weapon (scroll wheel or "1" "2" "3" buttons) that's not being updated in the game, so when you switch your weapon, the other players don't get updated about what weapon you currently have in the game, and that obviously is not correct.
You don't want your players to have to be just playing the wrong model at any point.
I used punRPC to fix that.