Initial commit
This manager allows us to start match finishes and set new matches and everything like that. What i used is something called events in photon networking. And events are basically ways of sending out a whole variety of different kinds of information at any particular time and we send the information out frome one of our players playing the game and then that information can be picked up by the other ones and something can be done with that information, basically.
So you'll  have way to send the information and ways to receive the information as well and it's all going to be handled with events, like I said.