Setting up Match Manager Systems
So to be able to do with these events, I need to have different information that our match manager is going to keep track of. So we're going to have our match manager keeping track of inforamtion about our players, for example, information about how many kills and how many deaths they've had. And we're also going to have our match manager keep track of what state our game is, examples: Is the match currently ongoing? Has somebody won the match? Has the match been finished? etc.
First thing I'm going to do is create how the player information is stored, I going to add a very simple class that won't have any unity functions built into it.

This class is going to store some different things about our player, first of all is the name that will have a public interest for actor, so actor is going to be a number that the network refers to each individual player as, because when you connect to netwrok , you're essentially assigned a number within the room, like your player. And within the class I also going to have an int value for how many kills the players have made and how many deaths they've received, to store all of this I used a constructor. I'm going to create a reference to stored information, a public list of player info that we're going to call all players.
All of this is just information that we're storing, I actually doing anything with it, so what I need to is to serialize the class.  I'm going to add a index for keep track of what is our position in this list. To determine which kind of events we're sending I used an enum (a specific category of data) that will be all type of events we're using.