Managing Events
To making these events start to happen, we need to set up the framework that will allow them all to take place, to do that I need to add some "extra functions" into our system that will allow us to control the events that are happening.
Example: We've got a new player event, a list players event, and we've got to change that event.
What we want for each of those is a function that will send the event and a function that will receive information about the event.
The sending functions we're going to control when that happens. For example: when someone dies, will update the kill stats. When somebody joins the match, will send a new player, sends everybody else on a network, knows that the players here and the list players will be a full list of all the players that are in our system.
When we receive a new player information or list player updates, that's obviously those are going to be events that have been sent out. And we have our own event up here, which is what's going to listen for events coming in.